sys/stm32: Fix SD driver to not read past end of buffer.
The FIFO in HAL SD driver from STMicro would read past the end of
the input buffer, which would hard fault when reading past the u area
in CCM RAM. Now read as bytes instead of 32-bit words, and keep a
dataremaining variable to trigger attempts to read more than what
comes from config.DataLength.

We only do blocks, so kill comments suggesting otherwise.

main() now #if 0 or #if 1 toggles which process is debugged with gdb.

The scheduler process (proc 0) goes all the way until the longjmp()
call from swtch().
The init() process (proc 1) goes all the way to user space and makes
a syscall to exec().

Small bug fixes.

+++ discobsd/sys/stm32/stm32f4xx_hal_sd.h
@@ -144,11 +144,11 @@ typedef struct
 
   HAL_LockTypeDef              Lock;             /*!< SD locking object                   */
 
-  uint32_t                     *pTxBuffPtr;      /*!< Pointer to SD Tx transfer Buffer    */
+  uint8_t                      *pTxBuffPtr;      /*!< Pointer to SD Tx transfer Buffer    */
 
   uint32_t                     TxXferSize;       /*!< SD Tx Transfer size                 */
 
-  uint32_t                     *pRxBuffPtr;      /*!< Pointer to SD Rx transfer Buffer    */
+  uint8_t                      *pRxBuffPtr;      /*!< Pointer to SD Rx transfer Buffer    */
 
   uint32_t                     RxXferSize;       /*!< SD Rx Transfer size                 */
